pyproject.toml
[project]
name = "graph-rag-system"
version = "4.0.0"
description = "Graph RAG (LlamaIndex + Neo4j) vs Vector RAG (ChromaDB) comparison system"
requires-python = ">=3.10"

dependencies = [
  "fastapi",
  "uvicorn",
  "gradio",
  "requests",
  "python-dotenv",
  "openai",
  "numpy",
  "neo4j>=5.0",
  "chromadb>=0.4",
  "Wikipedia-API>=0.6",
  "beautifulsoup4",
  "llama-index-core",
  "llama-index-graph-stores-neo4j",
  "llama-index-llms-openai-like",
  "llama-index-embeddings-openai",
]

[project.optional-dependencies]
experiments = [
  "scikit-learn",
  "scipy",
  "pandas",
  "nltk",
  "spacy",
  "matplotlib",
  "sentence-transformers",
]

[project.scripts]
backend = "app.backend:main"
frontend = "app.frontend:main"

[tool.uv]
package = true

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
